@@ -3,14 +3,14 @@ plugins {
 }
 
 android {
-    compileSdk 31
+    compileSdk 32
 
     defaultConfig {
         applicationId "api.software.salehunter"
         minSdk 23
-        targetSdk 31
-        versionCode 6
-        versionName "0.5.3 Pre-Alpha"
+        targetSdk 32
+        versionCode 7
+        versionName "0.6.4 Pre-Alpha"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -32,7 +32,6 @@ android {
             mavenCentral() //for fb auth sdk
             google() //for google auth
         }
-
     }
 
     viewBinding {
@@ -49,9 +48,26 @@ dependencies {
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.1'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+
+
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+
+    //MVVM
+    def lifecycle_version = "2.3.1"
+    // ViewModel:
+    implementation "androidx.lifecycle:lifecycle-viewmodel:$lifecycle_version"
+    // Saved state module:
+    implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:$lifecycle_version"
+    // LiveData:
+    implementation "androidx.lifecycle:lifecycle-livedata:$lifecycle_version"
+    implementation "androidx.lifecycle:lifecycle-reactivestreams:$lifecycle_version" //LiveData - RxJava Converter
+    // Annotation processor Java8:
+    implementation "androidx.lifecycle:lifecycle-common-java8:$lifecycle_version"
+
+    // RxJava
+    implementation 'io.reactivex.rxjava2:rxjava:2.1.9'
 
     //Navigation Component
     implementation 'androidx.navigation:navigation-fragment:2.4.1'
@@ -67,6 +83,7 @@ dependencies {
     //Retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0' //Json Factory Converter GSON
+    implementation 'com.squareup.retrofit2:adapter-rxjava2:2.3.0' //RxJava Factory Converter
 
     //Glide
     implementation 'com.github.bumptech.glide:glide:4.12.0'
@@ -87,4 +104,10 @@ dependencies {
     // If you want to additionally use the CameraX Extensions library
     implementation "androidx.camera:camera-extensions:${camerax_version}"
 
+    //Google Maps
+    implementation 'com.google.android.gms:play-services-maps:18.0.2'
+    implementation 'com.google.android.gms:play-services-location:19.0.1'
+
+    //Hello Charts
+    implementation 'com.github.lecho:hellocharts-library:1.5.8@aar'
 }
