@@ -5,6 +5,9 @@ import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
 
 import android.text.Editable;
 import android.text.TextWatcher;
@@ -13,16 +16,18 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.AnimationUtils;
 
-import java.util.regex.Pattern;
-
 import api.software.salehunter.databinding.FragmentForgotPasswordBinding;
+import api.software.salehunter.model.BaseResponseModel;
+import api.software.salehunter.util.DialogsProvider;
 import api.software.salehunter.util.TextFieldValidator;
 import api.software.salehunter.view.activity.AccountSign;
 import api.software.salehunter.R;
-import api.software.salehunter.model.EmailVerificationModel;
+import api.software.salehunter.viewmodel.fragment.accountSign.ForgotPasswordViewModel;
 
 public class ForgotPasswordFragment extends Fragment {
     private FragmentForgotPasswordBinding vb;
+    private ForgotPasswordViewModel viewModel;
+    private NavController navController;
 
     public ForgotPasswordFragment() {
         // Required empty public constructor
@@ -54,6 +59,9 @@ public class ForgotPasswordFragment extends Fragment {
             ((AccountSign) getActivity()).setBackButton(true);
         }
 
+        navController = Navigation.findNavController(view);
+        viewModel = new ViewModelProvider(this).get(ForgotPasswordViewModel.class);
+
         vb.forgotPasswordEmail.requestFocus();
         vb.forgotPasswordEmail.clearFocus();
 
@@ -82,26 +90,49 @@ public class ForgotPasswordFragment extends Fragment {
             }
         });
 
-        vb.forgotPasswordButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                boolean validData = true;
-
-                if(vb.forgotPasswordEmail.getError()!=null || vb.forgotPasswordEmail.getEditText().getText().length()==0){
-                    vb.forgotPasswordEmail.requestFocus();
-                    vb.forgotPasswordEmail.startAnimation(AnimationUtils.loadAnimation(getContext(),R.anim.missing));
-                    validData = false;
-                }
-
-                if(validData) {
-                    EmailVerificationModel forgotPAsswordModel = new EmailVerificationModel();
-                    forgotPAsswordModel.setEmail(vb.forgotPasswordEmail.getEditText().getText().toString());
-
-                    ((AccountSign)getActivity()).sendEmailVerification(forgotPAsswordModel);
-                }
-
-            }
+        vb.forgotPasswordButton.setOnClickListener(button -> {
+            if(isDataValid()) sendEmailVerification();
         });
+    }
+
+    boolean isDataValid(){
+        boolean validData = true;
+
+        if(vb.forgotPasswordEmail.getError()!=null || vb.forgotPasswordEmail.getEditText().getText().length()==0){
+            vb.forgotPasswordEmail.requestFocus();
+            vb.forgotPasswordEmail.startAnimation(AnimationUtils.loadAnimation(getContext(),R.anim.fieldmissing));
+            validData = false;
+        }
+
+        return validData;
+    }
+
+    void sendEmailVerification(){
+        DialogsProvider.get(getActivity()).setLoading(true);
+        viewModel.sendEmailVerification(vb.forgotPasswordEmail.getEditText().getText().toString())
+                .observe(getViewLifecycleOwner(), response ->{
+                    DialogsProvider.get(getActivity()).setLoading(false);
+
+                    switch (response.code()){
+                        case BaseResponseModel.SUCCESSFUL_OPERATION:
+                            Bundle bundle = new Bundle();
+                            bundle.putString("email", vb.forgotPasswordEmail.getEditText().getText().toString());
+                            navController.navigate(R.id.action_forgotPasswordFragment_to_verificationCodeFragment,bundle);
+                            break;
+
+                        case BaseResponseModel.FAILED_NOT_FOUND:
+                            DialogsProvider.get(getActivity()).messageDialog("Email isn't registered", "you don't have an account\nplease sign up first");
+                            break;
+
+                        case BaseResponseModel.FAILED_REQUEST_FAILURE:
+                            DialogsProvider.get(getActivity()).messageDialog("Request Error", "Please Check your connection.");
+                            break;
+
+                        default:
+                            DialogsProvider.get(getActivity()).messageDialog("Server Error","Code: "+ response.code());
+
+                    }
+                });
     }
 
 
